Sprint 0 Meeting Docs
This is for Spring 0. Uploaded the folder containing the Sprint Planning and the Sprint Stand Up done on November 1st. This contains the planning for what we want to achieve in the next sprint, and a standup that let's everyone check on how other group members are doing, and what they want to accomplish.