Sprint 1 Meeting Docs
This is for Sprint 1. Uploaded a folder that contains the sprint planning and retro, as well as the standup done on November 13th. The planning contains the tasks we want to achieve by the next sprint, the standup lets everyone check on how other members are doing and what they want to accomplish, and the retro lets us reflect on how we could've done better in the previous sprint.